Fix Archtypes
