+++ character/archtype.go
@@ -3,28 +3,30 @@ package character
 import (
 	"fmt"
 	"strings"
+
+	"github.com/robinje/multi-user-dungeon/core"
 )
 
 // DisplayArchetypes logs the loaded archetypes for debugging purposes.
-func DisplayArchetypes(g *Game) {
+func DisplayArchetypes(g *core.Game) {
 	for key, archtype := range g.ArcheTypes {
-		Logger.Debug("Archetype", "name", key, "description", archtype.Description)
+		core.Logger.Debug("Archetype", "name", key, "description", archtype.Description)
 	}
 }
 
 // LoadArchetypes retrieves all archetypes from the DynamoDB table and stores them in the Server's ArcheTypes map.
-func LoadArchetypes(g *Game) error {
+func LoadArchetypes(g *core.Game) error {
 	g.Mutex.Lock()
 	defer g.Mutex.Unlock()
 
-	var archetypes []Archetype
+	var archetypes []core.Archetype
 	err := g.Database.Scan("archetypes", &archetypes)
 	if err != nil {
 		return fmt.Errorf("error scanning archetypes table: %w", err)
 	}
 
 	if g.ArcheTypes == nil {
-		g.ArcheTypes = make(map[string]*Archetype)
+		g.ArcheTypes = make(map[string]*core.Archetype)
 	}
 
 	for _, archetype := range archetypes {
@@ -46,22 +48,7 @@ func LoadArchetypes(g *Game) error {
 		archetypeCopy.Abilities = lowerAbilities
 
 		g.ArcheTypes[archetype.ArchetypeName] = &archetypeCopy
-		Logger.Debug("Loaded archetype", "name", archetype.ArchetypeName, "description", archetype.Description)
-	}
-
-	return nil
-}
-
-// StoreArchetypes stores all archetypes from the Server's ArcheTypes map into the DynamoDB table.
-func (g *Game) StoreArchetypes() error {
-
-	for _, archetype := range g.ArcheTypes {
-		err := g.Database.Put("archetypes", *archetype)
-		if err != nil {
-			return fmt.Errorf("error storing archetype %s: %w", archetype.ArchetypeName, err)
-		}
-
-		Logger.Debug("Stored archetype", "name", archetype.ArchetypeName)
+		core.Logger.Debug("Loaded archetype", "name", archetype.ArchetypeName, "description", archetype.Description)
 	}
 
 	return nil
